Changed algorithm for getting recommendations.
Instead of using compatibility-calculator-style compatibility with liked, ok, and disliked, only use liked and did not like. This allows us to use the number of recommendations in common to calculate a 90% confidence interval for the % chance of you liking a given recommendor's recommendations. Be pessimistic and sort by the low end of the confidence interval.

This algorithm has the benefit of giving less weight to recomendors with few animes in common without using an arbitrary cutoff. It is statistically sound, with no hackery or guesswork. Users that are new to anime will be able to get recommendations right away.

Code still needs much cleaning up.