Fixed performance problem where it would take a full second to connect.
When connecting using a hostname with TcpClient, if the hostname resolves to
both an IPv4 and an IPv6 address, it tries the IPv6 address first and only
tries the IPv4 address if it fails. Since the rec service listens on IPv4,
clients would try to connect to the IPv6 localhost, fail, and only then
use IPv4 localhost.

Also added some logging while I was there.