Optimized DB access in FreshenMalDatabase for ~2.5x-3.5x greater throughput.
List entry inserts are batched for each user and inserted all at once.
Animes are upserted no more than once per run.

The database portion of putting in a user with 880 entries took 7.5 seconds
with the previous code: 117 entries/second. The database portion of putting in
a user with 1819 entries about 2/3 of the way through a 20-user run took
5.3 seconds: 343 entries/second. 2.9x throughput.

How much this optimization matters depends on how low the delay between MAL
requests is set.