[Web, RecEngine] Added a fourth category in AnimeRecs output for anime the user has in common with the recommender and has completed or dropped but has not rated.
Implemented by cleaning up AnimeRecsRecSource/MalAnimeRecsRecSource recommenders
classify the recommendations in the results. Instead of Liked/NotLiked/All,
there is Liked/NotLiked/Inconclusive/NotInCommon/All. "Inconclusive" includes
any anime that is on the user's anime list that is a recommendation but did not
get classified as liked or not liked. The web output only treats completed and
dropped as inconclusive. Plan to watch, watching, and on hold are treated as if
they were not in common.