[Web] Made a simple AnimeRecs output view.
Most users probably aren't interested in knowing the details of who their best
match was and what anime they had in common. Plus, it is not uncommon for users
in the database to have several hundred anime completed which can make the
output quite large. The detailed view is still accessible if you specify
?detailedResults=true.