Changed recservice protocol to avoid timing out after 2 minutes on Windows.
Don't use a half-close to signal end-of-message to the server. The client is
in the FIN_WAIT_2 state after the half-close is acknowledged by the server.
Both Windows and Linux have a FIN_WAIT_2 timeout after being in the FIN_WAIT_2
state for a certain amount of time without reading any data. On Windows, it is
2 minutes. It does not happen on Linux when connecting to localhost.

See http://stackoverflow.com/questions/13466639/why-is-socket-receive-timing-out-on-a-half-closed-connection-when-timeout-is-set

Changed the protocol to prefix messages with the message length to fix this.

2 minutes is normally plenty of time for the server to return a response,
but training some rec sources like SVDPlusPlus can take around 30 minutes.