Fix for Crunchyroll change.
Crunchyroll changed so that age-restricted animes are not
shown in the list of shows unless you are logged in
with an account over the age restriction.

Changed the crunchyroll stream source to prompt
for a username and password, use those to log in,
and the use the cookies from logging in when
getting the page with the list of anime available.