First cut at porting animerecs to dart.
It's still in jquery style. Need to add restoring pub packages on build
and running dart2js on release builds.

Also, using the "exportable" library for JSON serialization takes
the generated js from 128 kB to 217 kB and doesn't do deserialization.
Should look into alternative libraries or handcode serialization to
Map<String, Object>.

When saving the .dart file, it needs to be copied to the bin folder
for the changes to take effect. This happens during a build, but would
be nice for that to happen automatically during development. pub serve
is no good for this purpose because it doesn't write to the file
system, instead keeping stuff in memory and serving over HTTP on a
different port from the Nancy app.