[UpdateStreams] Add command line option to get CR HTML locally.
This addresses Crunchyroll's use of Cloudflare's "I'm under attack" mode,
which redirects visitors to a page that attempts to determine if
the client is a real web browser. As long as Crunchyroll is using it,
you must log in to Crunchyroll in a web browser, go to
http://www.crunchyroll.com/videos/anime/alpha?group=all,
save the page's HTML, and pass --crunchyroll-local-html-file=<path>
on the command line.