Make MyMediaLite support conditional on targeting full .NET framework.
MyMediaLite and its dependencies, C5 and MathNet.Numerics, do not
yet have netstandard packages.

C5 has an open issue and a PR for it:
* https://github.com/sestoft/C5/issues/55
* https://github.com/sestoft/C5/pull/50

MathNet.Numerics has two open issues for it:
* https://github.com/mathnet/mathnet-numerics/issues/421
* https://github.com/mathnet/mathnet-numerics/issues/386

In both cases the maintainer seems interested in supporting
netstandard but wants to wait for the dust to settle a bit more.

I have contributed to packaging MyMediaLite so I can help it
target netstandard once C5 and MathNet.Numerics have a netstandard
target, assuming the MyMediaLite maintainer is ok with any changes
required.

This is not a big deal as animerecs does not really use MyMediaLite
other than as an undocumented recommendation algorithm accessible by
changing the URL.

Pull in the MedallionRandom package for the Shuffle method to use in
place of MyMediaLite's implementation.