Add .net core target for AnimeRecs.FreshenMalDatabase.
In order for Common.Logging to talk to NLog (to get log messages
from the MalApi library flowing), created a netstandard compatible
adapter, based off the code in Common.Logging. Needed two small changes
to NLogLoggerFactoryAdapter:

* Remove constructor that takes a
System.Collections.Specialized.NameValueCollection, as that type is
not present in netstandard.

* Use AppContext.BaseDirectory instead of
AppDomain.CurrentDomain.BaseDirectory.

AppContext is not present in .NET 4.5 but is in .NET 4.6, so update
all projects in solution to target .NET 4.6.2 for their full framework
target instead of .NET 4.5.2.

MalApi broke with the upgrade to NET 4.6.2. A package reference
to System.Net.Http 4.3.1 is required to fix it. Upgrade to MalApi
4.0.0-Beta2 for all projects in the solution that use it, which adds
System.Net.Http 4.3.1 as a dependency, fixing the breakage.

Upgrade from NLog 3.1.0.0 to NLog 5.0.0-beta09. NLog 5 is the
first version that supports netstandard. Only update FreshenMalDatabase
for now, will do other projects as I get to them.

Use Microsoft.Extensions.Configuration for configuration instead of
app.config. As a result, SlowMonkey is no longer needed.