Fix deadlock preventing program termination in recservice.
In Windows running on .net core, pressing ctrl+c would correctly
initation the shutdown procedure but deadlocked on the line:

Console.CancelKeyPress -= SigIntHandler;

Presumably in .net core Console.CancelKeyPress is locked for
modification while an event handler for it is being executed.

Don't bother unregistering the signal handlers. Just set a static
flag for if shutting down is in progress and only initiate shutdown
if it's not already in progress.