Fix bug in WhenAllCancelOnFirstException.
Because canceling other tasks when a fault occurs can
result in running their continuations synchronously, the continuation
for those other tasks can end up taking over the bookkeeping of the
aggregate task, resulting in the aggregate task ending up in a canceled
state instead of a faulted state. The lock does no good here because
the synchronous continuation runs on the same thread.

This manifested in UpdateStreams where one stream source would error
and the aggregate task would enter a canceled state instead of a faulted
state.

The fix is to do the cancellation of other tasks last, after marking the
bookkeeping as complete.