[UpdateStreams] Don't return multiple Amazon streams with the same name.
Multiple streams with the same name are almost certainly just different
seasons, with a season selector dropdown in the stream page.

The problem if multiple streams are returned is that the URL is
an episode URL instead of a series URL and the URL changes
every run. Space Dandy on Anime Strike has two streams for two seasons, both
"Space Dandy (Original Japanese Version)". Every run wipes out the mapped IDs
for Space Dandy because the CSV writer doesn't know which returned stream
goes with which line in the CSV.

With only one stream returned from the stream source, the CSV writer is
able to preserve the mapped ID.